+++ src/mapload_temporary.rs
@@ -835,7 +835,6 @@ fn load_datatable_into_scene<R: Read + Seek>(
 
                                     ents.push(scene.spawn((
                                         StaticInstances(instanced_renderer, mesh_tag),
-                                        EntityWorldId(data.world_id),
                                     )));
                                 }
                                 Err(e) => {