+++ CHANGELOG.md
@@ -29,6 +29,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 
 ## Unreleased / Rolling Release
 
+## 0.4.0 - 2024-02-18
+
 ### Added
 
 - Auto updater by @cohaereo